+++ cocos/platform/CCFileUtils.cpp
@@ -1122,7 +1122,7 @@ bool FileUtils::createDirectory(const std::string& path)
 
     // Create path recursively
     subpath = "";
-    for (int i = 0, size = dirs.size(); i < size; ++i)
+    for (size_t i = 0, size = dirs.size(); i < size; ++i)
     {
         subpath += dirs[i];
         dir = opendir(subpath.c_str());
